Move HypIoRegion to hypercalls submodule too.
